requirements.txt
fastapi
uvicorn
tensorflow #replace with 'tensorflow-macos' on MAC 
pillow
jinja2